Add content to body of webpage and create new dir for images
For page content, I added a decription of the recipe using
a h4 element. Set up ingredients with an unorder list (<ul>) element and
order list (<ol>) element for the steps on how to make the recipe.

Also, new dir was created for use of images such as the padthai.webp.

As a result, webpage has body content with basic structure and image.